+++ src/molecules/DatePicker/PeriodPicker.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import { DatePickerContextProvider, DatePickerContext } from './contex';
+import { DatePickerContextProvider, DatePickerContext } from './context';
 import { DatePickerInput } from './DatePickerInput';
 import { DatePickerMenu } from './DatePickerMenu';
 import { Avatar, Icon } from '../../index';
@@ -8,11 +8,12 @@ type Props = {
   /**
    * After a new date has been set
    */
-  onSelectDate?: (date?: string) => void;
+  onSelectDateFrom?: (date?: string) => void;
+  onSelectDateTo?: (date?: string) => void;
 
   period?: {
-    start: string;
-    end: string;
+    start?: string;
+    end?: string;
   };
 
   size?: 'compact' | 'default';
@@ -33,11 +34,29 @@ type Props = {
    * @default "Norwgian short name. ['Ma', 'Ti', 'On', 'To', 'Fr', 'Lø', 'Sø']"
    */
   dayLabels?: Array<string>;
+
+  /**
+   * Format "yyyy-mm-dd"
+   * @default undefined
+   */
+  maxDate?: string;
+
+  /**
+   * Format "yyyy-mm-dd"
+   * @default undefined
+   */
+  minDate?: string;
 };
 
 export const PeriodPicker = (props: Props) => {
   return (
-    <DatePickerContextProvider {...props} isPeriodPicker={true} period={props.period}>
+    <DatePickerContextProvider
+      {...props}
+      onSelectDateFrom={props.onSelectDateFrom}
+      onSelectDateTo={props.onSelectDateTo}
+      isPeriodPicker={true}
+      period={props.period}
+    >
       <DatePickerContext.Consumer>
         {(contextValue) => (
           <div className="telia-date-picker telia-date-picker__period" ref={contextValue.datePickerRef}>
