+++ src/molecules/DatePicker/context/useDatePicker.ts
@@ -1,0 +1,3 @@
+import { useContext } from 'react';
+import { DatePickerContext } from './provider';
+export const useDatePicker = () => useContext(DatePickerContext);
